requirements.txt
fastapi
uvicorn
pydantic
asyncpg
google-genai
python-dotenv
